Switch Viewer.DisplaySize from Vector2 to Point (float->int) and update the code as needed. Add WindowManager.ScreenChanged() to separate the display code.
git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@485 a9f3cfff-6c41-7446-b0ca-052bd761e6eb

+++ Source/RunActivity/Processes/RenderProcess.cs
@@ -123,7 +123,7 @@ namespace ORTS
 
 		public static void InitializeShadowMapLocations(Viewer3D viewer)
 		{
-			var ratio = viewer.DisplaySize.X / viewer.DisplaySize.Y;
+			var ratio = (float)viewer.DisplaySize.X / viewer.DisplaySize.Y;
 			var fov = 45.0f;
 			var n = (float)0.5;
 			var f = (float)viewer.Settings.ShadowMapDistance;
